% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_aggregate.R
\name{catchment_aggregate}
\alias{catchment_aggregate}
\title{Aggregate Floating Catchment Area Ratios}
\usage{
catchment_aggregate(from, to = from, id = "GEOID", value = "access",
  consumers = "population", to_id = id, to_consumers = consumers,
  map = NULL, original_from = TRUE, return_type = "original",
  verbose = FALSE)
}
\arguments{
\item{from}{A matrix-like object with IDs, values, and number of consumers, or a vector
of values. These are the lower-level entities to be aggregated up to entries in \code{to}.}

\item{to}{A matrix-like object with IDs and total consumers. These are higher-level
entities containing \code{from} entities.}

\item{id}{The column name of IDs in \code{from} or a vector of IDs.}

\item{value}{The column name of values in \code{from} or a vector of values to be aggregated.}

\item{consumers}{The column name of consumer totals in \code{from}, or a vector
of said totals of the same length as \code{from} IDs.}

\item{to_id}{The column name of IDs in \code{to}, or a vector if IDs.}

\item{to_consumers}{The column name of consumer totals in \code{to}, or a vector
of said totals of the same length as \code{to} IDs.}

\item{map}{A named list, with names as or corresponding to \code{to} IDs, and vectors
containing \code{from} IDs as entries.}

\item{original_from}{Logical indicating whether the \code{from} values are original (default).
If \code{from} values are original and \code{consumers} are specified, \code{from} values will be
multiplied by \code{consumers} before aggregation. Set this to \code{FALSE} if this has already been done
(such as if you set \code{return_type} to \code{"region"} in \code{\link{catchment_ratio}}).}

\item{return_type}{Determines the values that are returned: \code{"original"} (default) for an average
for each \code{to} ID based on the region sum of \code{consumers}, or specified \code{to_consumers}.
If a number, original scores will be multiplied by the number then averaged like \code{"original"}. If
\code{"region"}, the sum for each \code{to} ID is returned.}

\item{verbose}{Logical; if \code{TRUE}, will print a log of the aggregation process.}
}
\value{
A vector with an aggregate value (determined by \code{return_type})
for each entry in \code{to}.
}
\description{
Calculate aggregate sums or means, potentially weighted by number of consumers.
}
\examples{
# lower-level entries prepended with higher-level IDs (like GEOIDs)
higher <- data.frame(id = c(100, 110), pop = c(500, 500))
lower <- data.frame(
  id = paste0(higher$id, 1:5),
  value = 1:5 / 10,
  pop = c(50, 10, 100, 40, 30)
)
catchment_aggregate(
  lower, higher,
  consumers = "pop", id = "id", value = "value",
  to_id = "id", verbose = TRUE
)
}
