% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_network.R
\name{catchment_network}
\alias{catchment_network}
\title{Extract a network of consumers and providers}
\usage{
catchment_network(connections, from_start = NULL, to_start = NULL)
}
\arguments{
\item{connections}{A list- or matrix-like object with \code{"from"} and \code{"to"} entries, such as that
returned from \code{\link{catchment_connections}}.}

\item{from_start, to_start}{The ID of a \code{from} (consumer) or \code{to} (provider) in \code{connections} from
which to trace out a network.}
}
\value{
A subsetted version of \code{connections} containing only connections traceable from \code{from_start}
and/or \code{to_start}.
}
\description{
Extract an interconnected set of consumers and providers from a set of connections within catchment areas.
}
\examples{
pop <- simulate_catchments()
connections <- catchment_connections(
  pop$consumers, pop$providers,
  weight = 1, max_cost = 1,
)
catchment_network(connections, 1)
}
