% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_norway.R
\docType{data}
\name{catchment_norway}
\alias{catchment_norway}
\title{Norwegian catchments}
\format{A rda file}
\source{
\href{http://www.nve.no}{NVE}
}
\description{
Catchment shapefile convert into R format.
The original dataset is available at http://nedlasting.nve.no/gis/
}
\references{
the Norwegian Water Resources and Energy Directorate (NVE)
}
\keyword{datasets}

