% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_census_population.R
\name{download_census_population}
\alias{download_census_population}
\title{Download U.S. Census Population Data}
\usage{
download_census_population(dir, state, year = 2019,
  include_margins = FALSE, include_commutes = FALSE, counties = NULL,
  tracts = NULL, blockgroups = NULL, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{dir}{Directory in which to save the file(s).}

\item{state}{Name, abbreviation, or FIPS code of the state.}

\item{year}{4 digit year, between 2009 and the most recent year.}

\item{include_margins}{Logical; if \code{TRUE}, will save and include estimate margins or error, which is a
\code{data.frame} with the same dimensions as the estimates \code{data.frame}.}

\item{include_commutes}{Logical; if \code{TRUE}, will download the Longitudinal Employer-Household Dynamics (LEHD)
Origin-Destination Employment Statistics (LODES) data for the state and block groups within selected tracts.}

\item{counties, tracts, blockgroups}{A vector of counties, tracts, or block group GEOIDs within the specified state to
filter block groups for (block groups within the specified counties or tracts, or matching the specified block group
GEOID). Only one can be specified (lower levels overwrite higher levels). Defaults to all block groups.}

\item{overwrite}{Logical; if \code{TRUE}, will remove any existing files before downloading and saving new versions.}

\item{verbose}{Logical; if \code{FALSE}, will not print status messages.}
}
\value{
A list with at least an \code{estimates} entry, and optionally \code{margins} and/or \code{commutes}
entries. The \code{extimates} and \code{margins} entries are \code{data.frame}s with block groups in rows,
population variables in columns, and person counts in cells. The \code{commutes} entry is a sparse matrix with
home block groups in rows, work block groups in columns, and job counts in cells (all jobs, summed across blocks).
}
\description{
Download and load U.S. census American Community Survey 5-year summary files at the block group level,
optionally including home-work commute statistics.
}
\examples{
\dontrun{
download_census_population(".", "va")
}
}
