% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_locs}
\alias{example_locs}
\title{Example location data}
\format{
## `example_locs`
A data frame with 100 rows and 4 columns:
\describe{
  \item{label}{Unique identifier for each location}
  \item{x, y}{The longitude (x) and latitude (y) for each facility}
  \item{weight}{The initial weight used to estimate facility "attractiveness"}
}
}
\source{
Simulated for package examples.
}
\description{
Simulated locations meant to represent health facility locations in Kapiri
Mposhi district, Zambia.
}
\keyword{datasets}
