% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_pop}
\alias{example_pop}
\title{Load the example population raster}
\source{
<https://github.com/wpgp/wopr/>
}
\usage{
example_pop()
}
\value{
A [terra::SpatRaster] (CRS EPSG:4326).
}
\description{
Loads an example WorldPop population raster for Kapiri Mposhi district,
Zambia, bundled with the package as a GeoTIFF. A function is used (rather than
a lazy-loaded dataset) because [terra::SpatRaster] objects cannot be
serialized into `.rda` files.
}
\examples{
pop <- example_pop()
pop
}
