% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_shp}
\alias{example_shp}
\title{Example boundary}
\format{
## `example_shp`
An `sf` object with a single polygon feature (CRS EPSG:4326).
}
\source{
<https://malariaatlas.org/malariaatlas-r-package-for-accessing-data/>
}
\description{
An example boundary of Kapiri Mposhi district in Zambia.
}
\keyword{datasets}
