% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{print.catchment_data}
\alias{print.catchment_data}
\title{Print a catchment_data object}
\usage{
\method{print}{catchment_data}(x, ...)
}
\arguments{
\item{x}{A \code{catchment_data} object.}

\item{...}{Further arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a catchment_data object
}
