% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-functions.R
\name{print.catchment_fit}
\alias{print.catchment_fit}
\title{Print a catchment_fit object}
\usage{
\method{print}{catchment_fit}(x, ...)
}
\arguments{
\item{x}{A \code{catchment_fit} object.}

\item{...}{Further arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a catchment_fit object
}
