% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_coord}
\alias{shift_coord}
\title{Shift coordinates to nearest within a set of provided locations}
\usage{
shift_coord(org_coords, valid_coords)
}
\arguments{
\item{org_coords}{A two column dataframe containing the X and Y coordinates to be shifted. Must be [X,Y] order.}

\item{valid_coords}{A two column dataframe containing the X,Y coordinates used to reference.}
}
\value{
A two column dataframe (or matrix) with the updated X,Y coordinates.
}
\description{
Shift coordinates to nearest within a set of provided locations
}
