% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_catchments.R
\name{simulate_catchments}
\alias{simulate_catchments}
\title{Simulate Catchment Areas}
\usage{
simulate_catchments(n = 100, weight = 1, ..., consumers_range = c(10,
  5000), long_center = -99, long_spread = 1, lat_center = 40,
  lat_spread = 1, as_sf = TRUE)
}
\arguments{
\item{n}{Number of consumers to generate.}

\item{weight}{Weight method used on Euclidean distances between consumers and providers.}

\item{...}{Passes additional arguments to \code{\link{catchment_weight}}.}

\item{consumers_range}{A numeric vector with entries for minimal and maximal consumer counts
from which to sample.}

\item{long_center}{Longitude to randomly place consumers around.}

\item{long_spread}{Standard deviation of the consumer's longitude distribution.}

\item{lat_center}{Latitude to randomly place consumers around.}

\item{lat_spread}{Standard deviation of the consumer's latitude distribution.}

\item{as_sf}{Logical; if \code{FALSE}, consumer and provider data are returned in regular \code{data.frame}s.}
}
\value{
A list with \code{consumers}, \code{providers}, \code{cost}, and \code{weight} entries.
}
\description{
Generate a random population of consumers and providers.
}
\examples{
pop <- simulate_catchments()
if (require("leaflet", quiet = TRUE)) {
  leaflet(pop$providers) |>
    addCircles(
      radius = 1e5, color = "#666", stroke = FALSE
    ) |>
    addCircles(
      data = pop$consumers, label = ~ paste0("Consumers: ", count, ", Access: ", access)
    ) |>
    addCircles(
      color = "#000", label = ~ paste("Provider", id)
    )
}
}
