% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{summary.catchment_data}
\alias{summary.catchment_data}
\title{Summarise a catchment_data object}
\usage{
\method{summary}{catchment_data}(object, ...)
}
\arguments{
\item{object}{A \code{catchment_data} object.}

\item{...}{Further arguments (currently unused).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summarise a catchment_data object
}
