% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{build_mesh}
\alias{build_mesh}
\title{Create INLA mesh}
\usage{
build_mesh(pixel_locs, mesh.args = mesh.args, verbose = F, ...)
}
\arguments{
\item{pixel_locs}{A two-column matrix containing pixel coordinates.}

\item{mesh.args}{A list of arguments passed to [fmesher::fm_mesh_2d_inla()].}

\item{verbose}{TRUE/FALSE: print the mesh arguments used.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An `fm_mesh_2d` object (also of class `inla.mesh`).
}
\description{
Create INLA mesh
}
