% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{build_mesh}
\alias{build_mesh}
\title{Create INLA mesh}
\usage{
build_mesh(pixel_locs, mesh.args = mesh.args, verbose = F, ...)
}
\arguments{
\item{pixel_locs}{A two-column matrix containing pixel coordinates}

\item{mesh.args}{A list contains arugments from INLA::inla.mesh.2d()}
}
\value{

}
\description{
Create INLA mesh
}
