% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment-package.R
\docType{package}
\name{catchment-package}
\alias{catchment}
\alias{catchment-package}
\title{catchment: Calculate spatial access and availability metrics}
\description{
Offers a range of spatial access and availability metrics, including variants of 2- and 3-step floating catchment areas: 2SFCA (Luo & Wang, 2003) <doi:10.1068/b29120>, E2FSCA (Lou & Qi, 2009) <doi:10.1016/j.healthplace.2009.06.002>, KD2SFCA / G2SFCA (Dai, 2010) <doi: 10.1016/j.healthplace.2010.06.012>, and 3SFCA (Wan, Zou, & Sternberg, 2012) <doi:10.1080/13658816.2011.624987>.
}
\author{
\strong{Maintainer}: Micah Iserman \email{micah.iserman@gmail.com}

Authors:
\itemize{
  \item Biocomplexity Institute [copyright holder]
}

}
