% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment-package.R
\docType{package}
\name{catchment-package}
\alias{catchment}
\alias{catchment-package}
\title{catchment: Weighted Spatial Point Catchment Modelling}
\description{
Create gravity-style catchment model for spatial point data, using individual weights and rasterized gridded data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PATH-Global-Health/catchment}
  \item Report bugs at \url{https://github.com/PATH-Global-Health/catchment/issues}
}

}
\author{
\strong{Maintainer}: Justin Millar \email{jmillar@path.org}

}
\keyword{internal}
