% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-functions.R
\name{catchment_model}
\alias{catchment_model}
\title{Fit catchment model}
\usage{
catchment_model(dat, time = TRUE)
}
\arguments{
\item{dat}{A [catchment_data] object created by [prepare_data()].}

\item{time}{TRUE/FALSE Print optimisation run time?}
}
\value{
A list of class \code{catchment_fit} containing:
  \describe{
    \item{obj}{The TMB AD object.}
    \item{fit}{The \code{nlminb} optimisation result.}
    \item{sdr}{The [TMB::sdreport()] result, or \code{NULL} if it failed.}
    \item{data}{The \code{catchment_data} list passed in.}
  }
}
\description{
Fits the gravity-style catchment model and computes parameter uncertainty
via [TMB::sdreport()].  Convergence is checked automatically and warnings are
issued when the optimiser reports failure, the max gradient is large, or the
Hessian is not positive-definite.
}
