% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-functions.R
\name{catchment_model}
\alias{catchment_model}
\title{Fit catchment model}
\usage{
catchment_model(dat, time = T)
}
\arguments{
\item{dat}{A list created by prepare_data()}

\item{time}{TRUE/FALSE Print run time?}
}
\value{
A list from model output
}
\description{
Fit catchment model
}
