% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-functions.R
\name{catchment_populations}
\alias{catchment_populations}
\title{Get estimated catchment populations}
\usage{
catchment_populations(mod)
}
\arguments{
\item{mod}{Model object from catchment_model}
}
\value{
a named vector containing estimated catchment populations
}
\description{
Get estimated catchment populations
}
