% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_travel_surface.R
\name{create_travel_surface}
\alias{create_travel_surface}
\title{Create travel time surfaces}
\usage{
create_travel_surface(
  friction_surface,
  points,
  extent_file = NA,
  id_col = NA,
  x_col = "x",
  y_col = "y",
  clip_flag = FALSE,
  transition_matrix_exists_flag = FALSE,
  individual_surfaces = FALSE,
  output_dir = NA,
  parallel = TRUE,
  cores = floor(parallel::detectCores()/2),
  check_existing = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{friction_surface}{A raster containing the travel cost per pixel.}

\item{points}{A dataframe containing fields for LABEL, X, and Y for point-level features.}

\item{extent_file}{A raster or shapefile used to define extent of points and friction surface.}

\item{id_col}{A character string corresponding to the column in points dataframe for LABEL field.}

\item{x_col}{A character string corresponding to the column in points dataframe for x coordinate field.}

\item{y_col}{A character string corresponding to the column in points dataframe for y coordinate field.}

\item{clip_flag}{TRUE/FALSE: Should the friction surface extent be clipped?}

\item{transition_matrix_exists_flag}{TRUE/FALSE: Has the already been constructed, must be in output_dir.}

\item{individual_surfaces}{TRUE/FALSE: Should individual travel surface be created for each point?}

\item{output_dir}{A character string for output directory (rasters and intermediate outputs).}

\item{parallel}{TRUE/FALSE: Should individual surfaces be created using parallel processing?}

\item{cores}{an integer defining the number of CPU cores used for parallel processing}

\item{check_existing}{TRUE/FALSE: Should rasters with LABEL in the output_dir be ignored?}

\item{overwrite}{TRUE/FALSE: Should rasters in output_dir be overwritten?}

\item{...}{}
}
\description{
Create travel time surfaces
}
