% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pop}
\alias{example_pop}
\title{Example population raster}
\format{
## `example_pop`
A RasterLayer object
}
\source{
<https://github.com/wpgp/wopr/>
}
\usage{
example_pop
}
\description{
An example population raster from WorldPop for Kapiri Mposhi district in Zambia.
}
\keyword{datasets}
