% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_shp}
\alias{example_shp}
\title{Example shapefile}
\format{
## `example_shp`
A SpatialPolygons object with no features:
}
\source{
<https://malariaatlas.org/malariaatlas-r-package-for-accessing-data/>
}
\usage{
example_shp
}
\description{
An example boundary shapefile containing the boundary of Kapiri Mposhi district in Zambia.
}
\keyword{datasets}
