% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob-surface.R
\name{get_prob_raster}
\alias{get_prob_raster}
\title{Get access probability raster from catchment mode}
\usage{
get_prob_raster(mod, id_label)
}
\arguments{
\item{mod}{Catchment model object}

\item{id_label}{A character string containing the label ID for the specified location}
}
\value{
A RasterLayer object
}
\description{
Get access probability raster from catchment mode
}
