% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-functions.R
\name{make_model_object}
\alias{make_model_object}
\title{Constructing model object for catchment model}
\usage{
make_model_object(dat, family = "poisson")
}
\arguments{
\item{dat}{A list with class \code{catchment_data}.}

\item{family}{Likelihood family: \code{"poisson"} or \code{"nb"}.}
}
\value{
A TMB AD function object.
}
\description{
Constructing model object for catchment model
}
