% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-functions.R
\name{make_model_object}
\alias{make_model_object}
\title{Constructing model object for catchment model}
\usage{
make_model_object(dat)
}
\arguments{
\item{dat}{A list with class(catchment_data)}
}
\value{
A TMB list object
}
\description{
Constructing model object for catchment model
}
