% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_data}
\alias{prepare_data}
\title{Preparing input data from catchment model}
\usage{
prepare_data(
  prob_mat_init,
  pop_raster,
  location_data,
  id_col = "label",
  weight_col = "weight",
  x_col = "x",
  y_col = "y",
  mesh.args = NULL
)
}
\arguments{
\item{prob_mat_init}{A matrix object.}

\item{pop_raster}{A raster object/}

\item{location_data}{A dataframe object which contains point-level data, including coordinates, labels, and weights.}

\item{id_col}{A character for the column in location_data that contains the labels for individual point. This should be a unique identifier (no duplicated values).}

\item{weight_col}{A character for the column in location_data that contains the weights for each point.}

\item{x_col}{A character for the X column}

\item{y_col}{A character for the Y column}

\item{mesh.args}{A list of arguments passed to [build_mesh()].}
}
\value{
a list with class `catchment_data`.
}
\description{
Preparing input data from catchment model
}
