% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_matrix_from_folder.R
\name{travel_mat_from_folder}
\alias{travel_mat_from_folder}
\title{Title}
\usage{
travel_mat_from_folder(dir, reference = NA, sparse = TRUE, progress = TRUE)
}
\arguments{
\item{dir}{A character string for the location of individual travel time rasters (NOTE: this folder should ONLY contain .tif files for travel time rasters).}

\item{reference}{A raster files used to determine which pixels should be included (typically this is the population raster).}

\item{sparse}{TRUE/FALSE: Should pixels with NA values be remove? Almost always this should be TRUE.}

\item{progress}{TRUE/FALSE Show a progress bar?}
}
\value{
A matrix that is n_pixels by n_facilities.
}
\description{
Title
}
